update readme, pump up version
